requirements.txt
fastapi
uvicorn
pinecone-client
sentence-transformers
pydantic
python-multipart
PyPDF2
google-genai
python-dotenv
numpy